Conflicts:
	session/subscribers.php

+++ session/subscribers.php
@@ -115,9 +115,39 @@ if ($action == 'edit' && ($user->rights->agefodd->creer | $user->rights->agefodd
 			if (! empty($conf->global->AGF_USE_REAL_HOURS) && $agfsta->status_in_session !== 4) {
 
 				$agfssh = new Agefoddsessionstagiaireheures($db);
-				$agfssh->fetch_all_by_session(GETPOST('sessid', 'int'), GETPOST('stagerowid', 'int'));
+				$agfssh->fetch_all_by_session(GETPOST('sessid', 'int'), GETPOST('modstagid', 'int'));
 				foreach ( $agfssh->lines as $heures ) {
-					$heures->delete($user);
+				    if ($agfsta->status_in_session == 3 )
+				    {
+				        
+				        $TCal = array();
+				        if (!in_array($heures->fk_calendrier, $TCal))
+				        {
+				            $cal = new Agefodd_sesscalendar($db);
+				            $cal->fetch($heures->fk_calendrier);
+				            
+				            $TCal[$heures->fk_calendrier] = ($cal->heuref - $cal->heured) / 3600;
+				        }
+				        
+				        if (floatval($heures->heures) !== $TCal[$heures->fk_calendrier])
+				        {
+    				        $new_heures = new Agefoddsessionstagiaireheures($db);
+    				        
+    				        $new_heures->fk_stagiaire = $heures->fk_stagiaire;
+    				        $new_heures->nom_stagiaire = $heures->nom_stagiaire;
+    				        $new_heures->fk_user_author = $heures->fk_user_author;
+    				        $new_heures->fk_calendrier = $heures->fk_calendrier;
+    				        $new_heures->fk_session = $heures->fk_session;
+    				        $new_heures->heures = $TCal[$heures->fk_calendrier];
+    				        $new_heures->datec = $heures->datec;
+    				        
+    				        $res = $new_heures->create($user);
+    				        
+    				        if ($res) $heures->delete($user);
+    				        else setEventMessage($langs->trans('ErrUpdateHeures'), 'errors');
+				        }
+				    }
+					else $heures->delete($user);
 				}
 			}
 
@@ -416,9 +446,10 @@ if ($action == 'updatetraineestatus') {
 	if ($result < 0) {
 		setEventMessage($agf->error, 'errors');
 	} else {
+	    $statusinsession = GETPOST('statusinsession', 'int');
 		$stagiaires = new Agefodd_session_stagiaire($db);
 		$stagiaires->fk_session_agefodd = $agf->id;
-		$result = $stagiaires->update_status_by_soc($user, 1, 0, GETPOST('statusinsession', 'int'));
+		$result = $stagiaires->update_status_by_soc($user, 1, 0, $statusinsession);
 		if ($result < 0) {
 			setEventMessage($stagiaires->error, 'errors');
 		} else {
@@ -427,8 +458,39 @@ if ($action == 'updatetraineestatus') {
 				foreach ( $stagiaires->lines as $trainee ) {
 					$heures = new Agefoddsessionstagiaireheures($db);
 					$heures->fetch_all_by_session($agf->id, $trainee->id);
-					foreach ( $heures->lines as $creneaux ) {
-						$creneaux->delete($user);
+					if (in_array($statusinsession, array(5,6))) // non-présent ou annulé
+					{
+						foreach ( $heures->lines as $creneaux ) {
+							$creneaux->delete($user);
+						}
+					}
+					elseif ($statusinsession == 3) // présent
+					{
+					   foreach ( $heures->lines as $creneaux ) {
+					       $TCal = array();
+					       if (!in_array($creneaux->fk_calendrier, $TCal))
+					       {
+					            $cal = new Agefodd_sesscalendar($db);
+					            $cal->fetch($creneaux->fk_calendrier);
+					            
+					            $TCal[$creneaux->fk_calendrier] = ($cal->heuref - $cal->heured) / 3600;
+							}
+
+					        $new_heures = new Agefoddsessionstagiaireheures($db);
+					        
+					        $new_heures->fk_stagiaire = $creneaux->fk_stagiaire;
+					        $new_heures->nom_stagiaire = $creneaux->nom_stagiaire;
+					        $new_heures->fk_user_author = $creneaux->fk_user_author;
+					        $new_heures->fk_calendrier = $creneaux->fk_calendrier;
+					        $new_heures->fk_session = $creneaux->fk_session;
+					        $new_heures->heures = $TCal[$creneaux->fk_calendrier];
+					        $new_heures->datec = $creneaux->datec;
+					        
+					        $res = $new_heures->create($user);
+					        
+					        if ($res) $creneaux->delete($user);
+					        else setEventMessage($langs->trans('ErrUpdateHeures'), 'errors');
+					    }
 					}
 				}
 			}
@@ -659,8 +721,7 @@ if (! empty($id)) {
 			print '<div class="tabsAction"><input type="submit" class="butAction" value="' . $langs->trans('Save') . '">';
 			print '<a class="butAction" href="' . $_SERVER['PHP_SELF'] . '?action=edit&id=' . $id . '">' . $langs->trans('Cancel') . '</a></div>';
 			print '</form>';
-			print
-					"<script type=\"text/javascript\">
+			print "<script type=\"text/javascript\">
 					$(document).ready(function () {
 						$('.fillin').click(function(e){
                             e.preventDefault();
